feat: introduce processor to handle messages
